+++ src/misc.ts
@@ -1,8 +1,5 @@
 import firebase from 'firebase';
 import { createBrowserHistory } from 'history';
-import { applyMiddleware, createStore } from 'redux';
-import thunk from 'redux-thunk';
-import rootReducer, { Action, IState } from './reducers';
 
 export const noop: () => void = () => undefined;
 
@@ -11,10 +8,6 @@ export function sleep (ms: number) {
 }
 
 export const appHistory = createBrowserHistory();
-export const store = createStore<IState, Action, {}, {}>(
-  rootReducer,
-  applyMiddleware<IState, Action>(thunk),
-);
 
 export interface IRecord {
   createdAt: firebase.firestore.Timestamp;
