+++ src/screens/HomePage.tsx
@@ -11,8 +11,8 @@ import LoadingView from '../independents/LoadingView';
 import * as ErrorLogs from '../models/ErrorLogs';
 import * as Profiles from '../models/Profiles';
 import * as Rooms from '../models/Rooms';
+import { Dispatch, IState } from '../models/Store';
 import path, { RoomLink } from '../path';
-import { Dispatch, IState } from '../reducers';
 
 const LogoImageContainer = styled.div`
   text-align: center;
