+++ src/screens/RoomWritePage.tsx
@@ -12,7 +12,7 @@ import { debounce, noop } from '../misc';
 import * as ErrorLogs from '../models/ErrorLogs';
 import * as Profiles from '../models/Profiles';
 import * as Rooms from '../models/Rooms';
-import { Dispatch, IState } from '../reducers';
+import { Dispatch, IState } from '../models/Store';
 import NotFoundPage from './NotFoundPage';
 
 const EditorContainer = styled.div`
